contact_service/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
certifi==2024.8.30
click==8.1.7
cloudinary==1.41.0
Deprecated==1.2.15
fastapi==0.115.5
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
limits==3.13.0
packaging==24.2
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wrapt==1.17.0
